+++ collectors/switch.go
@@ -279,6 +279,9 @@ func (s *SwitchCollector) Collect(ch chan<- prometheus.Metric) {
 	ch <- prometheus.MustNewConstMetric(collectErrors, prometheus.GaugeValue, errors, s.collector)
 	ch <- prometheus.MustNewConstMetric(collecTimeouts, prometheus.GaugeValue, timeouts, s.collector)
 	ch <- prometheus.MustNewConstMetric(collectDuration, prometheus.GaugeValue, time.Since(collectTime).Seconds(), s.collector)
+	if strings.HasSuffix(s.collector, "-runonce") {
+		ch <- prometheus.MustNewConstMetric(lastExecution, prometheus.GaugeValue, float64(time.Now().Unix()), s.collector)
+	}
 }
 
 func (s *SwitchCollector) collect() ([]PerfQueryCounters, float64, float64) {